+++ src/zksolc/arguments.rs
@@ -75,6 +75,11 @@ pub struct Arguments {
     #[structopt(long = "solc")]
     pub solc: Option<String>,
 
+    /// The EVM version to generate LLL IR for.
+    /// See https://github.com/matter-labs/era-compiler-llvm-context/blob/main/src/evm_version.rs for reference.
+    #[structopt(long = "evm-version")]
+    pub evm_version: Option<String>,
+
     /// Specify addresses of deployable libraries. Syntax: `<libraryName>=<address> [, or whitespace] ...`.
     /// Addresses are interpreted as hexadecimal strings prefixed with `0x`.
     #[structopt(short = "l", long = "libraries")]
@@ -230,6 +235,12 @@ impl Arguments {
                 );
             }
 
+            if self.evm_version.is_some() {
+                anyhow::bail!(
+                    "`evm-version` is not used in Yul, LLVM IR and EraVM assembly modes."
+                );
+            }
+
             if self.force_evmla {
                 anyhow::bail!("EVM legacy assembly mode is not supported in Yul, LLVM IR and EraVM assembly modes.");
             }
@@ -281,6 +292,9 @@ impl Arguments {
             }
             if !self.libraries.is_empty() {
                 anyhow::bail!("Libraries must be passed via standard JSON input.");
+            }
+            if self.evm_version.is_some() {
+                anyhow::bail!("EVM version must be passed via standard JSON input.");
             }
 
             if self.output_directory.is_some() {