+++ src/evmla/assembly/instruction/codecopy.rs
@@ -6,22 +6,22 @@
 /// Translates the contract hash copying.
 ///
 pub fn contract_hash<'ctx, D>(
-    context: &mut compiler_llvm_context::EraVMContext<'ctx, D>,
+    context: &mut era_compiler_llvm_context::EraVMContext<'ctx, D>,
     offset: inkwell::values::IntValue<'ctx>,
     value: inkwell::values::IntValue<'ctx>,
 ) -> anyhow::Result<()>
 where
-    D: compiler_llvm_context::EraVMDependency + Clone,
+    D: era_compiler_llvm_context::EraVMDependency + Clone,
 {
     let offset = context.builder().build_int_add(
         offset,
         context.field_const(
-            (compiler_common::BYTE_LENGTH_X32 + compiler_common::BYTE_LENGTH_FIELD) as u64,
+            (era_compiler_common::BYTE_LENGTH_X32 + era_compiler_common::BYTE_LENGTH_FIELD) as u64,
         ),
         "datacopy_contract_hash_offset",
     );
 
-    compiler_llvm_context::eravm_evm_memory::store(context, offset, value)?;
+    era_compiler_llvm_context::eravm_evm_memory::store(context, offset, value)?;
 
     Ok(())
 }
@@ -30,14 +30,14 @@ where
 /// Translates the library marker copying.
 ///
 pub fn library_marker<D>(
-    context: &mut compiler_llvm_context::EraVMContext<D>,
+    context: &mut era_compiler_llvm_context::EraVMContext<D>,
     offset: u64,
     value: u64,
 ) -> anyhow::Result<()>
 where
-    D: compiler_llvm_context::EraVMDependency + Clone,
+    D: era_compiler_llvm_context::EraVMDependency + Clone,
 {
-    compiler_llvm_context::eravm_evm_memory::store_byte(
+    era_compiler_llvm_context::eravm_evm_memory::store_byte(
         context,
         context.field_const(offset),
         context.field_const(value),
@@ -50,23 +50,23 @@ where
 /// Translates the static data copying.
 ///
 pub fn static_data<'ctx, D>(
-    context: &mut compiler_llvm_context::EraVMContext<'ctx, D>,
+    context: &mut era_compiler_llvm_context::EraVMContext<'ctx, D>,
     destination: inkwell::values::IntValue<'ctx>,
     source: &str,
 ) -> anyhow::Result<()>
 where
-    D: compiler_llvm_context::EraVMDependency + Clone,
+    D: era_compiler_llvm_context::EraVMDependency + Clone,
 {
     let mut offset = 0;
     for (index, chunk) in source
         .chars()
         .collect::<Vec<char>>()
-        .chunks(compiler_common::BYTE_LENGTH_FIELD * 2)
+        .chunks(era_compiler_common::BYTE_LENGTH_FIELD * 2)
         .enumerate()
     {
         let mut value_string = chunk.iter().collect::<String>();
         value_string.push_str(
-            "0".repeat((compiler_common::BYTE_LENGTH_FIELD * 2) - chunk.len())
+            "0".repeat((era_compiler_common::BYTE_LENGTH_FIELD * 2) - chunk.len())
                 .as_str(),
         );
 
@@ -76,7 +76,7 @@ where
             format!("datacopy_destination_index_{index}").as_str(),
         );
         let datacopy_value = context.field_const_str_hex(value_string.as_str());
-        compiler_llvm_context::eravm_evm_memory::store(
+        era_compiler_llvm_context::eravm_evm_memory::store(
             context,
             datacopy_destination,
             datacopy_value,