+++ src/evmla/assembly/instruction/name.rs
@@ -364,7 +364,7 @@ pub enum Name {
         /// The called function name.
         name: String,
         /// The called function key.
-        entry_key: compiler_llvm_context::EraVMFunctionBlockKey,
+        entry_key: era_compiler_llvm_context::EraVMFunctionBlockKey,
         /// The stack state hash after return.
         stack_hash: md5::Digest,
         /// The input size.
@@ -372,7 +372,7 @@ pub enum Name {
         /// The output size.
         output_size: usize,
         /// The return address.
-        return_address: compiler_llvm_context::EraVMFunctionBlockKey,
+        return_address: era_compiler_llvm_context::EraVMFunctionBlockKey,
     },
     /// The recursive function return instruction.
     #[serde(skip)]