+++ src/project/contract/mod.rs
@@ -11,7 +11,7 @@ use serde::Deserialize;
 use serde::Serialize;
 use sha3::Digest;
 
-use compiler_llvm_context::EraVMWriteLLVM;
+use era_compiler_llvm_context::EraVMWriteLLVM;
 
 use crate::build::contract::Contract as ContractBuild;
 use crate::project::Project;
@@ -39,7 +39,7 @@ impl Contract {
     ///
     pub fn new(
         path: String,
-        source_hash: [u8; compiler_common::BYTE_LENGTH_FIELD],
+        source_hash: [u8; era_compiler_common::BYTE_LENGTH_FIELD],
         source_version: SolcVersion,
         ir: IR,
         metadata_json: Option<serde_json::Value>,
@@ -91,13 +91,13 @@ impl Contract {
     pub fn compile(
         mut self,
         project: Project,
-        optimizer_settings: compiler_llvm_context::OptimizerSettings,
+        optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
         is_system_mode: bool,
         include_metadata_hash: bool,
-        debug_config: Option<compiler_llvm_context::DebugConfig>,
+        debug_config: Option<era_compiler_llvm_context::DebugConfig>,
     ) -> anyhow::Result<ContractBuild> {
         let llvm = inkwell::context::Context::create();
-        let optimizer = compiler_llvm_context::Optimizer::new(optimizer_settings);
+        let optimizer = era_compiler_llvm_context::Optimizer::new(optimizer_settings);
 
         let version = project.version.clone();
         let identifier = self.identifier().to_owned();
@@ -110,7 +110,7 @@ impl Contract {
             optimizer.settings().to_owned(),
         );
         let metadata_json = serde_json::to_value(&metadata).expect("Always valid");
-        let metadata_hash: Option<[u8; compiler_common::BYTE_LENGTH_FIELD]> =
+        let metadata_hash: Option<[u8; era_compiler_common::BYTE_LENGTH_FIELD]> =
             if include_metadata_hash {
                 let metadata_string = serde_json::to_string(&metadata).expect("Always valid");
                 Some(sha3::Keccak256::digest(metadata_string.as_bytes()).into())
@@ -129,7 +129,7 @@ impl Contract {
                     .map_err(|error| anyhow::anyhow!(error.to_string()))?
             }
             IR::ZKASM(ref zkasm) => {
-                let build = compiler_llvm_context::eravm_build_assembly_text(
+                let build = era_compiler_llvm_context::eravm_build_assembly_text(
                     self.path.as_str(),
                     zkasm.source.as_str(),
                     metadata_hash,
@@ -145,7 +145,7 @@ impl Contract {
             }
             _ => llvm.create_module(self.path.as_str()),
         };
-        let mut context = compiler_llvm_context::EraVMContext::new(
+        let mut context = era_compiler_llvm_context::EraVMContext::new(
             &llvm,
             module,
             optimizer,
@@ -153,14 +153,15 @@ impl Contract {
             include_metadata_hash,
             debug_config,
         );
-        context.set_solidity_data(compiler_llvm_context::EraVMContextSolidityData::default());
+        context.set_solidity_data(era_compiler_llvm_context::EraVMContextSolidityData::default());
         match self.ir {
             IR::Yul(_) => {
-                let yul_data = compiler_llvm_context::EraVMContextYulData::new(is_system_mode);
+                let yul_data = era_compiler_llvm_context::EraVMContextYulData::new(is_system_mode);
                 context.set_yul_data(yul_data);
             }
             IR::EVMLA(_) => {
-                let evmla_data = compiler_llvm_context::EraVMContextEVMLAData::new(version.default);
+                let evmla_data =
+                    era_compiler_llvm_context::EraVMContextEVMLAData::new(version.default);
                 context.set_evmla_data(evmla_data);
             }
             IR::LLVMIR(_) => {}
@@ -205,16 +206,19 @@ impl Contract {
 
 impl<D> EraVMWriteLLVM<D> for Contract
 where
-    D: compiler_llvm_context::EraVMDependency + Clone,
+    D: era_compiler_llvm_context::EraVMDependency + Clone,
 {
     fn declare(
         &mut self,
-        context: &mut compiler_llvm_context::EraVMContext<D>,
+        context: &mut era_compiler_llvm_context::EraVMContext<D>,
     ) -> anyhow::Result<()> {
         self.ir.declare(context)
     }
 
-    fn into_llvm(self, context: &mut compiler_llvm_context::EraVMContext<D>) -> anyhow::Result<()> {
+    fn into_llvm(
+        self,
+        context: &mut era_compiler_llvm_context::EraVMContext<D>,
+    ) -> anyhow::Result<()> {
         self.ir.into_llvm(context)
     }
 }