+++ src/yul/parser/statement/block.rs
@@ -136,11 +136,14 @@ impl Block {
     }
 }
 
-impl<D> compiler_llvm_context::EraVMWriteLLVM<D> for Block
+impl<D> era_compiler_llvm_context::EraVMWriteLLVM<D> for Block
 where
-    D: compiler_llvm_context::EraVMDependency + Clone,
+    D: era_compiler_llvm_context::EraVMDependency + Clone,
 {
-    fn into_llvm(self, context: &mut compiler_llvm_context::EraVMContext<D>) -> anyhow::Result<()> {
+    fn into_llvm(
+        self,
+        context: &mut era_compiler_llvm_context::EraVMContext<D>,
+    ) -> anyhow::Result<()> {
         let current_function = context.current_function().borrow().name().to_owned();
         let current_block = context.basic_block();