+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # The `zksolc` changelog
 
+## [Unreleased]
+
+### Added
+
+- An option to disable the system request memoization
+
 ## [1.3.23] - 2024-01-30
 
 ### Added